% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{append_time}
\alias{append_time}
\title{Add time as a variable to a database}
\usage{
append_time(x)
}
\arguments{
\item{x}{tibble database}
}
\value{
tibble database with time added
}
\description{
Add time as a variable to a database
}
