% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_cal_matrix.R
\name{commutation_matrix}
\alias{commutation_matrix}
\title{Generate the commutation matrix of size \eqn{p^2 \times p^2}}
\usage{
commutation_matrix(p)
}
\arguments{
\item{p}{Positive integer specifying the dimension of the square matrix.}
}
\value{
A \eqn{p^2 \times p^2} commutation matrix.
}
\description{
Constructs the commutation matrix \eqn{K_p}, which is a
\eqn{p^2 \times p^2} permutation matrix satisfying
\eqn{\mathrm{vec}(A^\top) = K_p \mathrm{vec}(A)}
for any \eqn{p \times p} matrix \eqn{A}.
}
\examples{
K2 <- commutation_matrix(2)
print(K2)

}
