% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeBaseNetBenefit.R
\name{computeBaseNetBenefit}
\alias{computeBaseNetBenefit}
\title{Compute base net benefit}
\usage{
computeBaseNetBenefit(
  y,
  xstart = 0.01,
  xstop = 0.99,
  step = 0.01,
  type = "treated"
)
}
\arguments{
\item{y}{outcome as integer}

\item{xstart}{start threshold}

\item{xstop}{stop threshold}

\item{step}{step threshold}

\item{type}{type of net benefit}
}
\value{
A data frame with the computed base net benefit
}
\description{
Compute base net benefit
}
