% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeCalibration.R
\name{computeCalibration}
\alias{computeCalibration}
\title{Compute calibration}
\usage{
computeCalibration(labs, preds, event_level = "1", scale = F, plot = T)
}
\arguments{
\item{labs}{Labels}

\item{preds}{Predictions}

\item{event_level}{Define event level}

\item{scale}{Scale calibration (default is FALSE)}

\item{plot}{Plot calibration (default is TRUE)}
}
\value{
Computed fraction of calibration
}
\description{
Compute calibration
}
