% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlcv_function.R
\name{computeCvAUC}
\alias{computeCvAUC}
\title{Compute Cross Validation Area Under the Curve}
\usage{
computeCvAUC(folds, model, confidence = 0.95)
}
\arguments{
\item{folds}{Nested tibble object with the training and test predictions, trn_preds and tst_preds respectively.}

\item{model}{Model}

\item{confidence}{Confidence}
}
\value{
Tibble with the computed AUC with CI for both training and test folds.
}
\description{
Estimate the area under the curve (AUC) metric from computed cross validation metrics.
}
