% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeNetBenefit.R
\name{computeNetBenefit}
\alias{computeNetBenefit}
\title{Compute net benefit}
\usage{
computeNetBenefit(pt, y, pred, type = "treated")
}
\arguments{
\item{pt}{probability threshold values}

\item{y}{outcome as integer}

\item{pred}{predictions (probabilities)}

\item{type}{type of benefit plot}
}
\value{
A data frame with the copmuted net benefit
}
\description{
Compute net benefit
}
