% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countPerc.R
\name{countPerc}
\alias{countPerc}
\title{Count and percentage}
\usage{
countPerc(x, var, value = NULL)
}
\arguments{
\item{x}{data fame}

\item{var}{variable}

\item{value}{value}
}
\value{

}
\description{
Simple function to count the frequencies of a variable (var) within a data frame (x) along with the percentage.
When a value (value) is provided, the function will return the number of defined observations and compute the percentage.
}
