% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{current_month}
\alias{current_month}
\title{Retrieve the current month}
\usage{
current_month(x = Sys.Date(), offset = 0)
}
\arguments{
\item{x}{Date object}

\item{offset}{numeric, days to offset the current month}
}
\value{
Date class object
}
\description{
Retrieve the current month
}
