% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cws.R
\name{cws}
\alias{cws}
\title{Clean Work Space}
\usage{
cws(pattern = "DIR", restart = F)
}
\arguments{
\item{pattern}{pattern of objects name to not be removed (default = "DIR")}

\item{restart}{restart R session (default = F)}
}
\description{
This function allows you to clear the global environment and remove all plots
}
\details{
Version 1.8 2020-10-16
}
