% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{decompose_filename}
\alias{decompose_filename}
\title{Given one or more filenames, decompose into a database}
\usage{
decompose_filename(
  x = c("/foo/bar/uvcomp_SCIMonthlyGlobal.u_wind.1981-09-01T000000tif",
    "/foo/bar/uvcomp_SCIDailyGlobal.v_wind.1981-09-01T000000.tif",
    "/foo/bar/stress_SCIDailyGlobal.x_tau.1981-09-01T180000.tif")
)
}
\arguments{
\item{x}{char vector of filenames in pattern 'source.param.date.ext'}
}
\value{
tibble database composed of the following fields
\itemize{
\item{source char the datasetid}
\item{date} Date
\item{hour char in \%H\%M\%S (HHMMSS) format}
\item{per char period as in one of "sizh", "day", "month"}
\item{param char, u_wind, windspeed, etc}
\item{nrt logical TRUE is near real-time}
}
}
\description{
Given one or more filenames, decompose into a database
}
