% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlcv_function.R
\name{dlcvInner}
\alias{dlcvInner}
\title{Double Loop Cross Validation (DLCV) Inner}
\usage{
dlcvInner(x_inner, wf, hp_grid)
}
\arguments{
\item{x_inner}{Configuration of the inner loop folds}

\item{wf}{Workflow}

\item{hp_grid}{Hyperparameter grid}
}
\value{
Tibble with all computed predictions for the inner train and training folds.
}
\description{
Functionality to perform the DLCV inner loop.
}
\details{
Version 1.0 2022-02-15
}
