% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlcv_lr.R
\name{dlcvLr}
\alias{dlcvLr}
\title{Double Loop Cross Validation Linear Regression}
\usage{
dlcvLr(folds, rec)
}
\arguments{
\item{folds}{rsample object with either group V-fold or the standard V-fold cross validation folds.}

\item{rec}{recipes recipe used for training}
}
\value{
Tibble with k outer loop models, and training and testing predictions.
}
\description{
Functionality to perform Linear Regression in a Double Loop Cross Validation (DLCV)
}
