% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlcv_function.R
\name{dlcvOuter}
\alias{dlcvOuter}
\title{Double Loop Cross Validation (DLCV) Outer}
\usage{
dlcvOuter(x_outer, tuned_model)
}
\arguments{
\item{x_outer}{Configuration of the outer loop folds}

\item{tuned_model}{Tuned model}
}
\value{
Tibble with all computed predictions for the outer train and training folds.
}
\description{
Functionality to perform the DLCV outer loop.
}
\details{
Version 1.0 2022-02-15
}
