% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlcv_rf.R
\name{dlcvRf}
\alias{dlcvRf}
\title{DLCV Random Forest With Hyperparameter Optimization}
\usage{
dlcvRf(folds, rec)
}
\arguments{
\item{folds}{rsample object with either group V-fold or the standard V-fold cross validation folds.}

\item{rec}{recipes recipe used for training}
}
\value{
Tibble with k outer loop models, and training and testing predictions.
}
\description{
DLCV Random Forest With Hyperparameter Optimization
}
