% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlcv_rf.R
\name{dlcvRfNoHp}
\alias{dlcvRfNoHp}
\title{DLCV Random Forest No Hyperparameter Optimization}
\usage{
dlcvRfNoHp(folds, rec)
}
\arguments{
\item{folds}{rsample object with either group V-fold or the standard V-fold cross validation folds.}

\item{rec}{recipes recipe used for training}
}
\value{
Tibble with k outer loop models, and training and testing predictions.
}
\description{
DLCV Random Forest No Hyperparameter Optimization
}
