% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erddap.R
\name{ed_base_url}
\alias{ed_base_url}
\title{Retrieve the erddap base url (with trailing slash!)}
\usage{
ed_base_url()
}
\value{
base URL for erdapp server at CoastWatch central operations
}
\description{
Retrieve the erddap base url (with trailing slash!)
}
