% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erddap.R
\name{ed_dataset_ids}
\alias{ed_dataset_ids}
\title{Retrieve the known Blended Winds dataset ids}
\usage{
ed_dataset_ids()
}
\value{
character vector of Dataset Id values)
}
\description{
Retrieve the known Blended Winds dataset ids
}
