% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erddap.R
\name{ed_fetch}
\alias{ed_fetch}
\title{Fetch one or more ERDDAP resources}
\usage{
ed_fetch(..., path = nbs_path("world"), verbose = FALSE)
}
\arguments{
\item{...}{arguments for}

\item{path}{char, the output path}

\item{verbose}{logical, if TRUE output messages}
}
\description{
Fetch one or more ERDDAP resources
}
