% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erddap.R
\name{ed_format_time}
\alias{ed_format_time}
\title{Format time for erddap}
\usage{
ed_format_time(x = as.POSIXct(Sys.time(), tz = "UTC"))
}
\arguments{
\item{x}{time or Date object}
}
\value{
character time as YYYY-mm-ddTHH:MM:SSZ"
}
\description{
Format time for erddap
}
