% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erddap.R
\name{ed_get_datasets}
\alias{ed_get_datasets}
\title{Retrieve a table of access info for one or more griddap datasets}
\usage{
ed_get_datasets(what = c("all", ed_dataset_ids()), url = ed_base_url())
}
\arguments{
\item{what}{character vector of one or more datasets ids or "all"
to retrieve all of them}

\item{url}{the url for the server}
}
\description{
Retrieve a table of access info for one or more griddap datasets
}
