% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erddap.R
\name{ed_range}
\alias{ed_range}
\title{Retrieve the range for a dimension or variable}
\usage{
ed_range(x, what = "time")
}
\arguments{
\item{x}{info class object - see \code{\link[rerddap]{info}}}

\item{what}{char the name of one dimension or variable}
}
\value{
two element vector of min and max
}
\description{
Retrieve the range for a dimension or variable
}
