% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erddap.R
\name{ed_retrieve}
\alias{ed_retrieve}
\title{Fetch a griddap resource}
\usage{
ed_retrieve(
  id = "noaacwBlendedWindsMonthly",
  fields = NULL,
  bb = NULL,
  times = NULL,
  cleanup = NULL
)
}
\arguments{
\item{id}{char specifies one product such as "noaacwBlendedWindsMonthly" (u, v and speed)}

\item{fields}{one or more character field names}

\item{bb}{named numeric or bbox class provides spatial bounding box}

\item{times}{two element Date or POSIXct vector with start and stop}

\item{cleanup}{logical or NULL.  If TRUE then always cleanup temporary files.
If FALSE never clean up temporary files (which will be cleaned when you
gracefully exix R).  If NULL then only cleanup temporary files if the
result does not inherit from "stars_proxy".}
}
\value{
stars object
}
\description{
Fetch a griddap resource
}
