% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erddap.R
\name{ed_time_origin}
\alias{ed_time_origin}
\title{Retrieve the time origin}
\usage{
ed_time_origin(x)
}
\arguments{
\item{x}{info class object - see \code{\link[rerddap]{info}}}
}
\value{
POSIXct origin
}
\description{
Retrieve the time origin
}
