% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emNMV_equicorrelation.R
\name{emNMV_equicorrelation}
\alias{emNMV_equicorrelation}
\title{Estimation under Equicorrelation Structure}
\usage{
emNMV_equicorrelation(y)
}
\arguments{
\item{y}{A numeric matrix of size \eqn{n \times p}, where each row
corresponds to an observation and each column to a variable.}
}
\value{
A list with two elements:
\item{mu_est}{The estimated mean vector \eqn{\mu}.}
\item{sigma_est}{The estimated covariance matrix \eqn{\Sigma} with equicorrelation structure.}
}
\description{
This function computes the maximum likelihood estimators (MLE)
of the mean vector \eqn{\mu} and the covariance matrix \eqn{\Sigma}
of a multivariate normal distribution, under the constraint that
\eqn{\Sigma} has an equicorrelation structure.
}
\details{
The covariance matrix is assumed to have the form:
\deqn{\Sigma = \sigma^2 \left[ (1-\rho) I_p + \rho J_p \right],}
where \eqn{I_p} is the identity matrix, \eqn{J_p} is the all-ones matrix,
\eqn{\rho} is the correlation parameter, and \eqn{\sigma^2} is a variance parameter.

The estimators are obtained by maximizing the likelihood under this restriction.
}
\examples{
set.seed(123)
Sigma <- 2 * ((1 - 0.5) * diag(3) + 0.5 * matrix(1, 3, 3))
y <- MASS::mvrnorm(50, mu = c(1,2,3), Sigma = Sigma)
result <- emNMV_equicorrelation(y)
result$mu_est
result$sigma_est

}
