% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emNMV_fixedmu.R
\name{emNMV_fixedmu}
\alias{emNMV_fixedmu}
\title{Maximum Likelihood Estimator of Covariance with Fixed Mean}
\usage{
emNMV_fixedmu(y, mu_0)
}
\arguments{
\item{y}{A numeric matrix of size \eqn{n \times p}, where each row
corresponds to an observation and each column to a variable.}

\item{mu_0}{A numeric vector of length \eqn{p}, the fixed mean vector.}
}
\value{
A list with two elements:
\item{mu_est}{The fixed mean vector \eqn{\mu_0}.}
\item{sigma_est}{A \eqn{p \times p} covariance matrix, the estimated covariance matrix.}
}
\description{
This function computes the maximum likelihood estimator (MLE) of
the covariance matrix \eqn{\Sigma} for a multivariate normal distribution,
assuming that the mean vector \eqn{\mu} is fixed at a given value \eqn{\mu_0}.
}
\details{
The covariance is estimated as the average of the outer products of the data
centered at the fixed mean \eqn{\mu_0}.
}
\examples{
set.seed(123)
y <- MASS::mvrnorm(50, mu = c(1,2), Sigma = matrix(c(1,0.3,0.3,1),2,2))
mu0 <- c(1,2)
result <- emNMV_fixedmu(y, mu0)
result$mu_est
result$sigma_est

}
