% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch_nbs}
\alias{fetch_nbs}
\title{Fetch NBS data using THREDDS/OPeNDAP services (preferred)}
\usage{
fetch_nbs(
  uri = NULL,
  ...,
  params = "all",
  bb = c(xmin = 0, ymin = -90, xmax = 360, ymax = 90),
  path = nbs_path("world"),
  verbose = FALSE
)
}
\arguments{
\item{uri}{char, one or more OPeNDAP URLs to data to fetch or the output
of \code{\link{query_nbs}}}

\item{...}{arguments passed to \code{\link{query_nbs}} if \code{code} is NULL}

\item{params}{one or more parameters such as "windspeed" o r"x_tau" to download
(or "all" to get all available for the given product)}

\item{bb}{numeric, named 4 element numeric of xmin, xmax, ymin, ymax or
a \code{\link[sf]{st_bbox}} class object.}

\item{path}{output path where to write the data}

\item{verbose}{logical, if TRUE output helpful messages}
}
\value{
tibble database
}
\description{
Data are fetched and written to disk.  You can provide either URLs for
opendap data or parameters to establish a search.
}
