% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findTodo.R
\name{findTodo}
\alias{findTodo}
\title{Find todos}
\usage{
findTodo(dir, save.output = F)
}
\arguments{
\item{dir}{= Main path of project}

\item{save.output}{= Boolean for saving the TODO output}
}
\description{
Find all todo annotatations in project directory
}
\details{
Version 1.0 2020-05-22
}
\keyword{findTodo}
