% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlcv_boruta.R
\name{fipBoruta}
\alias{fipBoruta}
\title{Feature importance Boruta DLCV}
\usage{
fipBoruta(x, plot = T, min_n = 0, y_nudge = NULL)
}
\arguments{
\item{x}{DLCV boruta object generated by the dlcvBoruta function}

\item{plot}{Boolean to visualise plot}

\item{min_n}{Minimal number of folds to visualise}

\item{y_nudge}{Nudge y coordinate to visualise number of times features was selected in fold.}
}
\value{
Tibble with the feature importances and number of times selected in folds.
}
\description{
Visualize feature importance of the Boruta DLCV results
}
