% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlcv_lr.R
\name{fipLr}
\alias{fipLr}
\title{Feature importance Linear Regression}
\usage{
fipLr(x, plot = T)
}
\arguments{
\item{x}{DLCV linear regression object generated by the dlcvLr function}

\item{plot}{Boolean to visualise plot}
}
\value{
Tibble with the feature importances in all folds.
}
\description{
Visualize feature importance of the Linear Regression DLCV results
}
