% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_information.R
\name{fisher_information}
\alias{fisher_information}
\title{Fisher Information Matrix for Multivariate Normal Birnbaum-Saunders Distribution}
\usage{
fisher_information(mu, sigma, eta)
}
\arguments{
\item{mu}{Mean vector of length \code{p}.}

\item{sigma}{Covariance matrix of dimension \code{p} x \code{p}, must be symmetric and positive definite.}

\item{eta}{Positive scalar shape parameter.}
}
\value{
A numeric matrix of dimension \code{p + p*(p+1)/2 + 1} by \code{p + p*(p+1)/2 + 1}, with attribute \code{"valid"} indicating if the matrix is symmetric and positive definite.
}
\description{
Computes the expected Fisher Information Matrix for the multivariate normal Birnbaum-Saunders distribution.
Some expectations are approximated using Monte Carlo integration.
}
\examples{
set.seed(1)
p <- 2
mu <- rep(0,p)
sigma <- diag(p)
eta <- 1
fisher_information(mu, sigma, eta)

}
