% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_information_NMV.R
\name{fisher_information_NMV}
\alias{fisher_information_NMV}
\title{Fisher Information Matrix for the Multivariate Normal}
\usage{
fisher_information_NMV(mu, sigma)
}
\arguments{
\item{mu}{A numeric vector of length p (mean vector).}

\item{sigma}{A symmetric, positive-definite covariance matrix of dimension p x p.}
}
\value{
A numeric matrix of dimension \code{p + p*(p+1)/2} by \code{p + p*(p+1)/2 + 1}, with attribute \code{"valid"} indicating if the matrix is symmetric and positive definite.
}
\description{
Computes the Fisher information matrix of a multivariate normal distribution
with mean vector `mu` and covariance matrix `sigma`.
}
