% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_model_tibble.R
\docType{data}
\name{multi_model_tibble}
\alias{multi_model_tibble}
\title{Multi Model Tibble}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 200 rows and 3 columns.
}
\usage{
multi_model_tibble
}
\description{
Multi Model Tibble
}
\keyword{datasets}
