% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs-package.R
\docType{package}
\name{nbs}
\alias{nbs-package}
\alias{nbs}
\title{Access NOAA NCEI Blended Seawinds (NBS v2)}
\description{
A small R package to access NOAA NCEI Blended Seawinds (NBS v2).
}
\seealso{
\href{https://coastwatch.noaa.gov/cwn/products/noaa-ncei-blended-seawinds-nbs-v2.html}{NOAA NCEI Blended Seawinds (NBS v2)} from NCEI
}
