% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs2.R
\name{NBS2}
\alias{NBS2}
\title{R6 class for accessing NBS v2 OpenDAP}
\description{
R6 class for accessing NBS v2 OpenDAP
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{product}}{char, the product name (see \link{nbs_tally})}

\item{\code{base_uri}}{char, the base uri (see \link{nbs_base_url})}

\item{\code{NC}}{ncdf4 object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NBS2-new}{\code{NBS2$new()}}
\item \href{#method-NBS2-finalize}{\code{NBS2$finalize()}}
\item \href{#method-NBS2-build_uri}{\code{NBS2$build_uri()}}
\item \href{#method-NBS2-close_nc}{\code{NBS2$close_nc()}}
\item \href{#method-NBS2-open_nc}{\code{NBS2$open_nc()}}
\item \href{#method-NBS2-get_res}{\code{NBS2$get_res()}}
\item \href{#method-NBS2-get_lon}{\code{NBS2$get_lon()}}
\item \href{#method-NBS2-get_lat}{\code{NBS2$get_lat()}}
\item \href{#method-NBS2-get_epoch}{\code{NBS2$get_epoch()}}
\item \href{#method-NBS2-get_time}{\code{NBS2$get_time()}}
\item \href{#method-NBS2-get_varnames}{\code{NBS2$get_varnames()}}
\item \href{#method-NBS2-generate_database}{\code{NBS2$generate_database()}}
\item \href{#method-NBS2-get_nav}{\code{NBS2$get_nav()}}
\item \href{#method-NBS2-get_var}{\code{NBS2$get_var()}}
\item \href{#method-NBS2-clone}{\code{NBS2$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NBS2-new"></a>}}
\if{latex}{\out{\hypertarget{method-NBS2-new}{}}}
\subsection{Method \code{new()}}{
Create a new NBS2 object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NBS2$new(product = "uvcomp_SCIMonthlyGlobal", base_uri = nbs_base_url())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{product}}{char, the product name (see \link{nbs_tally})}

\item{\code{base_uri}}{char, the base uri (see \link{nbs_base_url})}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{NBS2} object with an opened ncdf4 object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NBS2-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-NBS2-finalize}{}}}
\subsection{Method \code{finalize()}}{
Cleanup a NBS2 object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NBS2$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NBS2-build_uri"></a>}}
\if{latex}{\out{\hypertarget{method-NBS2-build_uri}{}}}
\subsection{Method \code{build_uri()}}{
build a URL for the resource
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NBS2$build_uri()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NBS2-close_nc"></a>}}
\if{latex}{\out{\hypertarget{method-NBS2-close_nc}{}}}
\subsection{Method \code{close_nc()}}{
close the ncdf4 object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NBS2$close_nc()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NBS2-open_nc"></a>}}
\if{latex}{\out{\hypertarget{method-NBS2-open_nc}{}}}
\subsection{Method \code{open_nc()}}{
open the ncdf4 connection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NBS2$open_nc()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NBS2-get_res"></a>}}
\if{latex}{\out{\hypertarget{method-NBS2-get_res}{}}}
\subsection{Method \code{get_res()}}{
retrieve the spatial resolution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NBS2$get_res()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
two element numeric vector of (res_x, res_y)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NBS2-get_lon"></a>}}
\if{latex}{\out{\hypertarget{method-NBS2-get_lon}{}}}
\subsection{Method \code{get_lon()}}{
retrieve a vector of longitudes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NBS2$get_lon()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
numeric vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NBS2-get_lat"></a>}}
\if{latex}{\out{\hypertarget{method-NBS2-get_lat}{}}}
\subsection{Method \code{get_lat()}}{
retrieve a vector of latitudes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NBS2$get_lat()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
numeric vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NBS2-get_epoch"></a>}}
\if{latex}{\out{\hypertarget{method-NBS2-get_epoch}{}}}
\subsection{Method \code{get_epoch()}}{
retrieve the time origin
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NBS2$get_epoch()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Date or POSIXct class element
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NBS2-get_time"></a>}}
\if{latex}{\out{\hypertarget{method-NBS2-get_time}{}}}
\subsection{Method \code{get_time()}}{
retrieve the time vector
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NBS2$get_time()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Date or POSIXct class vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NBS2-get_varnames"></a>}}
\if{latex}{\out{\hypertarget{method-NBS2-get_varnames}{}}}
\subsection{Method \code{get_varnames()}}{
retrieve variable names
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NBS2$get_varnames(exclude = c("crs", "mask"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{exclude}}{NULL or char, one or mor evariables to exclude}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
charcater vector of available variables
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NBS2-generate_database"></a>}}
\if{latex}{\out{\hypertarget{method-NBS2-generate_database}{}}}
\subsection{Method \code{generate_database()}}{
generate a likely database
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NBS2$generate_database()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
tibble database
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NBS2-get_nav"></a>}}
\if{latex}{\out{\hypertarget{method-NBS2-get_nav}{}}}
\subsection{Method \code{get_nav()}}{
retrieve a navigation list for extracting rasters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NBS2$get_nav(
  bb = c(xmin = 288, ymin = 39, xmax = 297, ymax = 46),
  varid = "u_wind",
  time = 1
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{bb}}{4 element vector of (xmin, ymin, xmax, ymax) or an sf-like object
from which a bbox can be extracted}

\item{\code{varid}}{char, the name of the variable to extract}

\item{\code{time}}{numeric, Date or POSIXct single time to extract}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NBS2-get_var"></a>}}
\if{latex}{\out{\hypertarget{method-NBS2-get_var}{}}}
\subsection{Method \code{get_var()}}{
Extract an array or stars oject
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NBS2$get_var(
  time = 1,
  bb = c(xmin = 288, ymin = 39, xmax = 297, ymax = 46),
  varid = "u_wind",
  nav = NULL,
  form = c("stars", "array")[1]
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{time}}{numeric, Date or POSIXct single time to extract}

\item{\code{bb}}{4 element vector of (xmin, ymin, xmax, ymax) or an sf-like object
from which a bbox can be extracted}

\item{\code{varid}}{char, the name of the variable to extract}

\item{\code{nav}}{NULL or list, if not a list compute the navigation list as needed}

\item{\code{form}}{char, one of 'array' or 'stars' (default) specifying the output type}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
stars or array object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NBS2-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NBS2-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NBS2$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
