% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs2.R
\name{NBS2_get_nav}
\alias{NBS2_get_nav}
\title{Compute a navigation list given a bounding box}
\usage{
NBS2_get_nav(
  X,
  bb = c(xmin = 288, ymin = 39, xmax = 297, ymax = 46),
  varid = "u_wind",
  time = 1
)
}
\arguments{
\item{X}{NBS2 class object}

\item{bb}{4 element vector of (xmin, ymin, xmax, ymax) or an sf-like object
from which a bbox can be extracted}

\item{varid}{char, the name of the variable to extract}

\item{time}{index, Date or POSIXct single time}
}
\description{
Compute a navigation list given a bounding box
}
