% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs.R
\name{NBS2_get_var}
\alias{NBS2_get_var}
\title{Extract an array or stars oject}
\usage{
NBS2_get_var(
  X,
  time = 1,
  bb = c(288, 39, 297, 46),
  varid = "u_wind",
  nav = NULL,
  form = c("stars", "array")[1]
)
}
\arguments{
\item{X}{NBS2 class object}

\item{time}{numeric, Date or POSIXct single time to extract}

\item{bb}{4 element vector of (xmin, ymin, xmax, ymax) or an sf-like object
from which a bbox can be extracted}

\item{varid}{char, the name of the variable to extract}

\item{nav}{NULL or list, if not a list compute the navigation list as needed}

\item{form}{char, one of 'array' or 'stars' (default) specifying the output type}
}
\value{
stars or array object
}
\description{
Extract an array or stars oject
}
