% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs.R
\name{nbs_generate_database}
\alias{nbs_generate_database}
\title{Generate a database-let from the the filename of the of a ncdf4 object}
\usage{
nbs_generate_database(x)
}
\arguments{
\item{x}{ncdf4 object}
}
\value{
tibble database composed of the following fields
\itemize{
\item{source char the datasetid}
\item{date} Date
\item{hour char in \%H\%M\%S (HHMMSS) format}
\item{per char period as in one of "sizh", "day", "month"}
\item{param char, u_wind, windspeed, etc}
\item{nrt logical TRUE is near real-time}
}
}
\description{
Generate a database-let from the the filename of the of a ncdf4 object
}
