% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs.R
\name{nbs_get_var}
\alias{nbs_get_var}
\title{Extract an array or stars oject}
\usage{
nbs_get_var(
  x,
  time = 1,
  bb = c(xmin = 288, ymin = 39, xmax = 297, ymax = 46),
  varid = "u_wind",
  nav = NULL,
  form = c("stars", "array")[1],
  verbose = FALSE
)
}
\arguments{
\item{x}{ncdf4 object}

\item{time}{numeric, Date or POSIXct single time to extract}

\item{bb}{4 element vector of (xmin, ymin, xmax, ymax) or an sf-like object
from which a bbox can be extracted}

\item{varid}{char, the name of the variable to extract. One or more varid
may be specified for a single time.}

\item{nav}{NULL or list, if not a list compute the navigation list as needed}

\item{form}{char, one of 'array' or 'stars' (default) specifying the output type}

\item{verbose}{logical, if TRUE output messages}
}
\value{
stars or array object
}
\description{
Extract an array or stars oject
}
