% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{nbs_path}
\alias{nbs_path}
\title{Retrieve a path}
\usage{
nbs_path(..., root = root_path())
}
\arguments{
\item{...}{file segements to add to the root path. See \code{\link[base]{file.path}}}

\item{root}{the root path to the NBS data}
}
\value{
character path description
}
\description{
Retrieve a path
}
