% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs.R
\name{nbs_product}
\alias{nbs_product}
\title{Retrieve a named character vector of paths relative to the root path.
Each element is named with the product name}
\usage{
nbs_product(x = NULL)
}
\arguments{
\item{x}{missing or character, if missing return all products. If character,
return only those requested}
}
\value{
named character vector
}
\description{
Retrieve a named character vector of paths relative to the root path.
Each element is named with the product name
}
