% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs.R
\name{nbs_tally}
\alias{nbs_tally}
\title{Provides a brief tally/overview of available data}
\usage{
nbs_tally()
}
\value{
a tibble of summary information
}
\description{
Provides a brief tally/overview of available data
}
