% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{nbs_thredds_url}
\alias{nbs_thredds_url}
\title{Base THREDDS url}
\usage{
nbs_thredds_url(data = FALSE)
}
\arguments{
\item{data}{logical, if TRUE return the thredds database url otherwise
the plain vanilla coastwatch url}
}
\description{
Base THREDDS url
}
