% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs.R
\name{nbs_vars}
\alias{nbs_vars}
\title{Retrieve nbs variables}
\usage{
nbs_vars(
  x,
  drop = c("time_run", "time_offset", "crs", "mask", "latitude", "longitude", "zlev")
)
}
\arguments{
\item{x}{ncdf4 object}

\item{drop}{character items to not list as variables.  Set to 'none' to drop none.}
}
\value{
character vector
}
\description{
Retrieve nbs variables
}
