% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs.R
\name{nbs_which_time}
\alias{nbs_which_time}
\title{Convert a real time to an index}
\usage{
nbs_which_time(x, time = 1)
}
\arguments{
\item{x}{ncdf4 object}

\item{time}{numeric index, Date or POSIXct}
}
\value{
index
}
\description{
Convert a real time to an index
}
