% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objSize.R
\name{objSize}
\alias{objSize}
\title{Object Size}
\usage{
objSize(objs)
}
\arguments{
\item{objs}{character vector with object name(s)}
}
\value{
A tibble containing all objects arranged on object size (mb).
}
\description{
Compute memory size of objects in the R environment.
}
\details{
Version 1.0 2022-02-15
}
