% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erddap.R
\name{parse_range}
\alias{parse_range}
\title{Parse a min-max string}
\usage{
parse_range(x = "0, 1")
}
\arguments{
\item{x}{character in the form of "min, max"}
}
\value{
two element numeric range
}
\description{
Parse a min-max string
}
