% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCalibration.R
\name{plotCalibration}
\alias{plotCalibration}
\title{Plot calibration}
\usage{
plotCalibration(x)
}
\arguments{
\item{x}{tibble with the computed mean_pred and obs_frac per model}
}
\value{
A ggplot containing the calibration plot
}
\description{
Plot calibration
}
