% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlcv_function.R
\name{plotDisMet}
\alias{plotDisMet}
\title{Plot Discriminative Metrics}
\usage{
plotDisMet(x, bl = NA, group = "id", title = "", legend = F, breaks = F)
}
\arguments{
\item{x}{Nested tibble object}

\item{bl}{Discriminative metrics for a baseline model (sensitivity, specificity and PPV)}

\item{group}{Group parameter}

\item{title}{Plot title}

\item{legend}{Boolean to visualise legend}

\item{breaks}{Boolean to plot additional breaks and lines}
}
\value{
Cowplot visualisation with four plots; ROC and PRC curves for both the training and tests folds.
}
\description{
Functionality to plot both ROC and PRC curves for both the training and tests folds.
}
\details{
Version 1.0 2022-02-15
}
