% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPrc.R
\name{plotPrc}
\alias{plotPrc}
\title{Plot Precision Recall Curve}
\usage{
plotPrc(x, group = NULL, title = NULL)
}
\arguments{
\item{x}{Tibble with precision and recall computed by the performance R package}

\item{group}{Variable to group coordinates}

\item{title}{Plot title}
}
\value{
A ggplot containing the precision recall curve
}
\description{
Plot Precision Recall Curve
}
