% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_nbs}
\alias{query_nbs}
\title{Query the THREDDS catalog}
\usage{
query_nbs(
  period = c("six-hour", "daily", "monthly")[2],
  dates = c(as.Date("1987-07-09"), as.Date("1987-09-09")),
  product = c("uvcomp", "stress")[1],
  nrt = FALSE,
  base_uri = nbs_thredds_url()
)
}
\arguments{
\item{period}{chr, one of "six-hour", "daily" (default) or "monthly"}

\item{dates}{Date, start and stop dates to query}

\item{product}{char, one of "uvcomp" (default) or "stress"}

\item{nrt}{logical, if TRUE query for near real time (default is false)}

\item{base_uri}{char, the base uri for the thredds catalog}
}
\value{
a tibble with zero or more resources identified
}
\description{
Query the THREDDS catalog
}
