% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_nbs}
\alias{read_nbs}
\title{Read one or more NBS files as stars}
\usage{
read_nbs(x, path)
}
\arguments{
\item{x}{tibble database of the data elements to read as rasters}

\item{path}{char, the path to the data storage}
}
\value{
stars object
}
\description{
Read one or more NBS files as stars
}
