% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erddap.R
\name{read_noaacwBlendedWind}
\alias{read_noaacwBlendedWind}
\title{Read in a table of griddap products}
\usage{
read_noaacwBlendedWind(
  filename = system.file("erddap/noaacwBlendedWind.csv", package = "nbs")
)
}
\arguments{
\item{filename}{char, the name of the file to read}
}
\value{
tibble
}
\description{
Read in a table of griddap products
}
\seealso{
\href{https://coastwatch.noaa.gov//erddap/search/index.html?page=1&itemsPerPage=1000&searchFor=noaacwBlendedWind}{Coastwatch listing page}
}
