% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replacePattern.R
\name{replacePattern}
\alias{replacePattern}
\title{Replace Pattern}
\usage{
replacePattern(
  dir,
  file.types = c("r", "R", "Rmd"),
  file.pattern = NULL,
  pattern,
  replacement
)
}
\arguments{
\item{dir}{= directory}

\item{file.types}{= File types that should be assessed in dir}

\item{file.pattern}{= Search pattern to identify relevant files}

\item{pattern}{= Pattern to be replaced in all relevant file types}

\item{replacement}{= Replacement to replace files that match pattern}
}
\description{
This function searches documents for matches to the argument pattern and
replaces in with a new pattern
}
\details{
Version 1.2 2019-11-19
}
