% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_model_tibble.R
\docType{data}
\name{single_model_tibble}
\alias{single_model_tibble}
\title{Single Model Tibble}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 100 rows and 2 columns.
}
\usage{
single_model_tibble
}
\description{
Single Model Tibble
}
\keyword{datasets}
