% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb.R
\name{stars_to180}
\alias{stars_to180}
\title{Translate a stars from \link{0,360} to \link{-180, 180}}
\usage{
stars_to180(x)
}
\arguments{
\item{x}{stars object}
}
\value{
a spatially ranslated version of the input \code{x}
}
\description{
Translate a stars from \link{0,360} to \link{-180, 180}
}
