% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb.R
\name{stars_to360}
\alias{stars_to360}
\title{Translate a stars from \link{-180, 180} to \link{0,360}}
\usage{
stars_to360(x)
}
\arguments{
\item{x}{stars object}
}
\value{
a spatially ranslated version of the input \code{x}
}
\description{
Translate a stars from \link{-180, 180} to \link{0,360}
}
