% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeFunc.R
\name{summarizeFunc}
\alias{summarizeFunc}
\title{Summarize Function}
\usage{
summarizeFunc(x, colname, check_normality = T, normality = T, print_plot = F)
}
\arguments{
\item{x}{Data (Data frame / Tibble)}

\item{colname}{Name of column of interest (character)}

\item{check_normality}{Check normality (Boolean)}

\item{normality}{Normality, if known (Boolean)}

\item{print_plot}{Print plot (Boolean)}
}
\value{

}
\description{
Summarize Function
}
\examples{
mtcars \%>\% with_groups(am, ~ .x \%>\% summarizeFunc("hp"))
}
