% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb.R
\name{to180}
\alias{to180}
\title{Convert \link{0,360} longitudes to \link{-180, 180}}
\usage{
to180(x)
}
\arguments{
\item{x}{numeric vector, no check is done for being withing -180, 180 range}
}
\value{
numeric vector
}
\description{
Convert \link{0,360} longitudes to \link{-180, 180}
}
