% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb.R
\name{to360}
\alias{to360}
\title{Convert \link{-180,180} longitudes to \link{0,360}}
\usage{
to360(x)
}
\arguments{
\item{x}{numeric vector, no check is done for being withing 0,360 range}
}
\value{
numeric vector
}
\description{
Convert \link{-180,180} longitudes to \link{0,360}
}
