% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{write_database}
\alias{write_database}
\title{Write a database}
\usage{
write_database(x, path, filename = "database.csv.gz", drop = "time")
}
\arguments{
\item{x}{tibble database}

\item{path}{char, path description to the database}

\item{filename}{char, the filename to save to}

\item{drop}{char one or more column names to drop or "none" to keep them all}
}
\value{
the input database
}
\description{
Write a database
}
