% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bb.R
\name{bb_to180}
\alias{bb_to180}
\title{Convert bounding box \link{0,360} longitudes to \link{-180, 180}}
\usage{
bb_to180(x)
}
\arguments{
\item{x}{numeric bounding box vector, no check is done for being withing -180, 180 range}
}
\value{
numeric bounding box vector
}
\description{
Bounding boxes are 4 element vectors of \link{left, right, bottom, top}
}
