% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_cal_matrix.R
\name{commutation_matrix}
\alias{commutation_matrix}
\title{Generate the commutation matrix of size p^2 x p^2}
\usage{
commutation_matrix(p)
}
\arguments{
\item{p}{Positive integer specifying the dimension of the square matrix.}
}
\value{
A \(p^2 \times p^2\) commutation matrix.
}
\description{
Constructs the commutation matrix \(K_p\) which is a \(p^2 \times p^2\) permutation matrix
satisfying \(\text{vec}(A^T) = K_p \text{vec}(A)\) for any \(p \times p\) matrix \(A\).
}
\examples{
K2 <- commutation_matrix(2)
print(K2)

}
