% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dlcv_regularization.R
\name{dlcvL1}
\alias{dlcvL1}
\title{Double Loop Cross Validation L1}
\usage{
dlcvL1(folds, rec)
}
\arguments{
\item{folds}{rsample object with either group V-fold  or the standard V-fold cross validation folds}

\item{rec}{recipes recipe used for training}
}
\value{

}
\description{
Double Loop Cross Validation with features selected in the inner loop with either one of the following functions:
\itemize{
\item dlcvInnerL1Tidy
\item dlcvInnerL1Glmnet
}
}
