% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbs.R
\name{nbs_time_units}
\alias{nbs_time_units}
\alias{nbs_epoch}
\alias{nbs_time}
\alias{nbs_trange}
\alias{nbs_tcount}
\title{Working with time contents in NBS ncdf4 objects}
\usage{
nbs_time_units(x)

nbs_epoch(x)

nbs_time(x)

nbs_trange(x)

nbs_tcount(x)
}
\arguments{
\item{x}{ncdf4 object}
}
\value{
nbs_time_units returns a character time unit (days or hours)

nbs_epoch returns a single Date or POSIXct starting point from which time is figured

nbs_time returns a Date or POSIXct vector

nbs_trange returns a 2-element vector of Date or POSIXct start and stop values

nbs_tcount returns the number of time steps
}
\description{
Time may be in days or hours since a common epoch (1978-01-01). \code{nms_time_unit()}
returns a string such as 'days' or 'hours'. \code{nbs_epoch()} returns a Date or POSIXct
class starting element. \code{nbs_get_time()} returns all of the time elements as Date
or POSIXct. \code{nbs_trange()} returns a two element vector of start/stop values as
Date or POSIXct class.  \code{nbs_tcount()} returns the number of time steps.
}
