% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCalibration.R
\name{plotCalibration}
\alias{plotCalibration}
\title{Plot calibration}
\usage{
plotCalibration(
  x,
  plot_errorbar = T,
  plot_smooth = T,
  group = NULL,
  title = NULL
)
}
\arguments{
\item{x}{tibble with the computed mean_pred and obs_frac per model}

\item{plot_errorbar}{Plot error bar (default = TRUE)}

\item{plot_smooth}{Plot smooth lines (default = TRUE)}

\item{group}{Variable to group coordinates}

\item{title}{Plot title}
}
\value{
A ggplot containing the calibration plot
}
\description{
Plot calibration
}
