% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRoc.R
\name{plotRoc}
\alias{plotRoc}
\title{Plot Receiver Operator Curve}
\usage{
plotRoc(x, group = NULL, title = NULL, breaks = F)
}
\arguments{
\item{x}{Tibble with sensitivity and specificity computed by the performance R package}

\item{group}{Variable to group coordinates}

\item{title}{Plot title}

\item{breaks}{Boolean to plot additional breaks and lines}
}
\value{
A ggplot containing the receiver operator curve
}
\description{
Plot Receiver Operator Curve
}
