% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRoc.R
\name{plotRoc}
\alias{plotRoc}
\title{Plot Receiver Operator Curve}
\usage{
plotRoc(x, colour = "Model")
}
\arguments{
\item{x}{Tibble with sensitivity and specificity computed by the performance R package}

\item{colour}{Name of colour}
}
\value{
A ggplot containing the receiver operator curve
}
\description{
Plot Receiver Operator Curve
}
